// src/app/models/vendors/raw-vendors-list-record.model.ts
export interface RawVendorsListRecord {
  attending_flag_description: string;
  attending_flag_id: number;
  booth_number_id: number;
  business_name: string;
  email: string;
  first_name: string;
  id: number;
  invite_create_time: string;
  invite_update_time: string;
  last_name: string;
  number: string;
  redeemed: number;
}